Linked groups with page content

+++ src/main/java/com/youthministry/dao/GroupDao.java
@@ -32,7 +32,7 @@ public interface GroupDao {
      *
      * @param  int Id
      */
-    public Group getGroupById(int id);
+    public Group getGroupById(Long id);
     
     /**
      * Get Event
